Added WiFi UI Files and updated robot code for orientation correction

+++ IISC_Lab/motors.h
@@ -12,6 +12,6 @@
 //Function Declarations
 void init_motors(void);
 void motor(uint8_t index, int value);
-void motors(int *value);
+void motors(float *value);
 void set_motor(uint8_t index, int value);
